# swagger-spec.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}}}},"/image/stories":{"post":{"operationId":"StoriesController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateStoryDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Story"}}}}}},"get":{"operationId":"StoriesController_findAll","parameters":[{"name":"populate","required":true,"in":"query","schema":{"default":[],"type":"array","items":{"type":"string"}}},{"required":true,"name":"pagesTags","in":"query","schema":{"default":{},"type":"object"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"object"}}}}}}}},"/image/stories/{uuid}":{"get":{"operationId":"StoriesController_findOne","parameters":[{"name":"uuid","required":true,"in":"path","schema":{"type":"string"}},{"name":"populate","required":true,"in":"query","schema":{"default":[],"type":"array","items":{"type":"string"}}},{"required":true,"name":"pagesTags","in":"query","schema":{"default":{},"type":"object"}}],"responses":{"200":{"description":""}}},"patch":{"operationId":"StoriesController_update","parameters":[{"name":"uuid","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateStoryDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Story"}}}}}},"delete":{"operationId":"StoriesController_remove","parameters":[{"name":"uuid","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":""}}}},"/image/chapters":{"post":{"operationId":"ChaptersController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateChapterDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Chapter"}}}}}},"get":{"operationId":"ChaptersController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Chapter"}}}}}}}},"/image/chapters/{uuid}":{"get":{"operationId":"ChaptersController_findOne","parameters":[{"name":"uuid","required":true,"in":"path","schema":{"type":"string"}},{"name":"populate","required":true,"in":"query","schema":{"default":[],"type":"array","items":{"type":"string"}}},{"required":true,"name":"pagesTags","in":"query","schema":{"default":{},"type":"object"}}],"responses":{"200":{"description":""}}},"patch":{"operationId":"ChaptersController_update","parameters":[{"name":"uuid","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateChapterDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Chapter"}}}}}},"delete":{"operationId":"ChaptersController_remove","parameters":[{"name":"uuid","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":""}}}},"/image/pages":{"post":{"operationId":"PagesController_create","parameters":[],"requestBody":{"required":true,"description":"A new page","content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/CreatePageDto"}}}},"responses":{"204":{"description":""}}},"get":{"operationId":"PagesController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PageDto"}}}}}}}},"/image/pages/{uuid}":{"get":{"operationId":"PagesController_findOne","parameters":[{"name":"uuid","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageDto"}}}}}},"patch":{"operationId":"PagesController_update","parameters":[{"name":"uuid","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdatePageDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageDto"}}}}}},"delete":{"operationId":"PagesController_remove","parameters":[{"name":"uuid","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":""}}}},"/files/{uuid}":{"get":{"operationId":"FilesController_findOne","parameters":[{"name":"uuid","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/File"}}}}}}}},"/files/{uuid}/download":{"get":{"operationId":"FilesController_findOneImage","parameters":[{"name":"uuid","required":true,"in":"path","schema":{"type":"string"}},{"name":"width","required":true,"in":"query","schema":{"type":"number"}},{"name":"height","required":true,"in":"query","schema":{"type":"number"}},{"name":"fit","required":true,"in":"query","schema":{"enum":["cover","contain","fill","inside","outside"],"type":"string"}}],"responses":{"200":{"description":""}}}}},"info":{"title":"Image server","description":"A server to organize and serve image","version":"0.0.1","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"CreateStoryDto":{"type":"object","properties":{"title":{"type":"string"},"mediaType":{"type":"string"}},"required":["title","mediaType"]},"Story":{"type":"object","properties":{}},"UpdateStoryDto":{"type":"object","properties":{"title":{"type":"string"},"mediaType":{"type":"string"}}},"CreateChapterDto":{"type":"object","properties":{"title":{"type":"string"},"number":{"type":"string"},"storyUuid":{"type":"string"}},"required":["title","number","storyUuid"]},"Chapter":{"type":"object","properties":{}},"UpdateChapterDto":{"type":"object","properties":{"title":{"type":"string"},"number":{"type":"string"},"storyUuid":{"type":"string"}}},"CreatePageDto":{"type":"object","properties":{"number":{"type":"string"},"storyUuid":{"type":"string"},"bookUuid":{"type":"string"},"chapterUuid":{"type":"string"},"file":{"type":"string","format":"binary"},"tags":{"type":"array","items":{"type":"string"}}},"required":["number","storyUuid","file"]},"PageDto":{"type":"object","properties":{"number":{"type":"string"},"fileUrl":{"type":"string"}},"required":["number","fileUrl"]},"UpdatePageDto":{"type":"object","properties":{"number":{"type":"string"},"storyUuid":{"type":"string"},"bookUuid":{"type":"string"},"chapterUuid":{"type":"string"},"file":{"type":"string","format":"binary"},"tags":{"type":"array","items":{"type":"string"}}}},"File":{"type":"object","properties":{}}}}}